requirements.txt
matplotlib==3.7.2
Pillow==11.0.0
scikit_learn==1.3.0
seaborn==0.13.2
torch==2.4.1+cu124
torchvision==0.19.1+cu124
